4.15, 4.16*, 4.17, 4.18, 4.19, 4.20* & 4.21*

+++ osa4/controllers/blogs.js
@@ -1,22 +1,35 @@
-const blogsRouter = require('express').Router()
+const blogsRouter = require("express").Router()
+const jwt = require("jsonwebtoken")
 // const blogsRouter = require('../models/blog')
-const Blog = require('../models/blog')
+const Blog = require("../models/blog")
+const User = require("../models/user")
 
+// const getTokenFrom = request => {
+//   const authorization = request.get("authorization")
+//   if (authorization && authorization.toLowerCase().startsWith("bearer")) {
+//     return authorization.substring(7)
+//   }
+//   return null
+// }
 
+// rami
+// eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InJhbWkiLCJpZCI6IjVkN2NlNTkyODE5ZGM3MTQxNDU4MzQ0NiIsImlhdCI6MTU2ODQ4NDUxM30.ZOjYZH6-IZpKMtCKCJG4MNCSZOIoq44CznKozuIv_CM
 
-blogsRouter.get('/', async (request, response, next) => {
-  const blogs = await Blog.find({})
+blogsRouter.get("/", async (request, response, next) => {
+  const blogs = await Blog.find({}).populate("user", { name: 1, username: 1 })
+
   try {
-    response.json(blogs.map(blog => {
-      return blog.toJSON()
-    }))
-  }
-  catch(error) {
+    response.json(
+      blogs.map(blog => {
+        return blog.toJSON()
+      })
+    )
+  } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.get('/:id', (request, response, next) => {
+blogsRouter.get("/:id", (request, response, next) => {
   Blog.findById(request.params.id)
     .then(blog => {
       if (blog) {
@@ -28,35 +41,65 @@ blogsRouter.get('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post("/", async (request, response, next) => {
   const body = request.body
-  if(!request.body.title || !request.body.author) {
+
+  const token = request.token
+
+  if (!request.body.title || !request.body.author) {
     return response.status(400).end()
   }
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    likes: body.likes || 0,
-    url: body.url
-  })
+
   try {
-  const savedBlog = await blog.save()
-  response.json(savedBlog.toJSON())
-  }
-  catch(error) {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      likes: body.likes || 0,
+      url: body.url,
+      user: user._id
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog.toJSON())
+  } catch (error) {
     next(error)
   }
 })
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then(() => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+blogsRouter.delete("/:id", async (request, response, next) => {
+  try {
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (decodedToken.id.toString() === blog.user.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      return response.status(204).end()
+    } else {
+      return response
+        .status(401)
+        .send({ error: "token did not match owner of blog" })
+    }
+  } catch (error) {
+    next(error)
+  }
 })
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put("/:id", (request, response, next) => {
   const body = request.body
 
   const blog = {
@@ -72,18 +115,5 @@ blogsRouter.put('/:id', (request, response, next) => {
     })
     .catch(error => next(error))
 })
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 module.exports = blogsRouter
